+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0 / 2025-08
+* Updated the plugin to use Akamai Edgegrid v11
+* Updated the plugin to use latest Prometheus Go libraries
+* Updated the plugin to use the latest Akamai Reporting API -- [authoritative-dns-traffic-by-time](https://techdocs.akamai.com/reporting/v1/reference/authoritative-dns-traffic-by-time)
+
+
 ## 0.1.0 / 2021-01-31
 
 Initial release.